refactor(nix): resolve models-dev-gen deps with bun2nix
The generator's two runtime dependencies were vendored by hand: two
pkgs.fetchurl calls with npm integrity strings copied out of the pinned
models-dev input's bun.lock, extracted into a flat node_modules with tar.
Bumping the input meant transcribing hashes between two lockfiles.

Resolve them through bun2nix instead, from a package.json that pins the
same remeda 2.33.7 and zod 3.24.2 that input's lockfile records. The
generated bun.nix carries hashes identical to the fetchurl calls it
replaces, and both snapshots the build emits are unchanged.

Also move the tool into nix/tools/models-dev-gen so its Nix expression,
the scripts it copies into the workspace and its dependency manifests sit
together: models-dev-pricing.nix becomes default.nix, models-dev-gen.ts
becomes gen.ts, and models-dev-compact.ts drops its now-redundant prefix.
The remaining references follow: the node test invocation in the justfile
and in the git hook, the tsconfig exclude and the oxlint ignore pattern.
