+++ nix/tools/models-dev-gen/bun.nix
@@ -1,0 +1,24 @@
+# Autogenerated by `bun2nix`, editing manually is not recommended
+#
+# Set of Bun packages to install
+#
+# Consume this with `fetchBunDeps` (recommended)
+# or `pkgs.callPackage` if you wish to handle
+# it manually.
+{
+  copyPathToStore,
+  fetchFromGitHub,
+  fetchgit,
+  fetchurl,
+  ...
+}:
+{
+  "remeda@2.33.7" = fetchurl {
+    url = "https://registry.npmjs.org/remeda/-/remeda-2.33.7.tgz";
+    hash = "sha512-cXlyjevWx5AcslOUEETG4o8XYi9UkoCXcJmj7XhPFVbla+ITuOBxv6ijBrmbeg+ZhzmDThkNdO+iXKUfrJep1w==";
+  };
+  "zod@3.24.2" = fetchurl {
+    url = "https://registry.npmjs.org/zod/-/zod-3.24.2.tgz";
+    hash = "sha512-lY7CDW43ECgW9u1TcT3IoXHflywfVqDYze4waEz812jR/bZ8FHDsl7pFQoSZTz5N+2NqRXs8GBwnAwo3ZNxqhQ==";
+  };
+}
