+++ nix/tools/models-dev-gen/compact.ts
